+++ web/app/scripts/compiled-es5/directives/calculate-percentage.js
@@ -37,6 +37,9 @@ _module.directive('calculatePercentage', [_calculatePercentage2.default.componen
 
             // Using true to compare the sub-elements.
             scope.$watch('model', function (answers) {
+                if (totalQuestionCount === 0) {
+                    return;
+                }
                 var numberOfCompletedQuestions = CalculateCompletion.countNumberOfAnsweredQuestions(answers);
                 scope.percentageComplete = CalculateCompletion.calculatePercentageComplete(numberOfCompletedQuestions, totalQuestionCount);
             }, true);