+++ web/app/scripts/compiled-es5/services/calculate-percentage.js
@@ -55,12 +55,18 @@ _module.service(moduleProperties.componentName, [function () {
     };
 
     this.calculatePercentageComplete = function (completedQuestions, totalQuestionCount) {
+        var zero = 0 + '%';
+
+        if (totalQuestionCount === 0) {
+            return zero;
+        }
+
         var result = completedQuestions / totalQuestionCount * 100;
         if (result >= 0) {
             return Math.round(result) + '%';
         }
 
-        return 0 + '%';
+        return zero;
     };
 }]);